Testing Update.
Apparently, even if a model isn't used the Shapes file still demands that there are entries in the data model file.  I guess this makes sense to some extent.

So an approach might be to have the RM Shapes as just a text file and only use (copy/paste) the ones you need for each data model.

Frustrating to say the least.